+++ src/features/NextFreeOffset.cpp
@@ -25,7 +25,7 @@ struct GroupIDSource final {
         if (auto eobj = typeinfo_cast<EffectGameObject*>(obj)) {
             used.insert(eobj->m_centerGroupID);
             used.insert(eobj->m_targetGroupID);
-            used.insert(eobj->m_activateGroup);
+            log::info("targetGroupID: {}", eobj->m_targetGroupID);
         }
     }
 };