+++ src/features/about/AboutBEButton.cpp
@@ -2,6 +2,7 @@
 #include <Geode/modify/EditorPauseLayer.hpp>
 #include <Geode/ui/BasedButtonSprite.hpp>
 #include "AboutBEPopup.hpp"
+#include "SupportPopup.hpp"
 
 using namespace geode::prelude;
 
@@ -10,6 +11,65 @@ class $modify(AboutBEPauseLayer, EditorPauseLayer) {
     bool init(LevelEditorLayer* lel) {
         if (!EditorPauseLayer::init(lel))
             return false;
+
+        if (!Mod::get()->template getSavedValue<bool>("dont-show-support-button")) {
+            if (auto topMenu = this->getChildByID("top-menu")) {
+                auto supportSpr = CCSprite::createWithSpriteFrameName("support-btn.png"_spr);
+
+                auto glow = CCSprite::createWithSpriteFrameName("support-btn-glow.png"_spr);
+                glow->setZOrder(2);
+                glow->setBlendFunc({ GL_SRC_ALPHA, GL_ONE });
+                glow->setOpacity(125);
+                glow->runAction(CCRepeat::create(
+                    CCSequence::create(
+                        CCEaseInOut::create(CCFadeTo::create(1.5f, 0), 4.f),
+                        CCEaseInOut::create(CCFadeTo::create(1.5f, 125), 4.f),
+                        nullptr
+                    ),
+                    1000
+                ));
+                supportSpr->addChildAtPosition(glow, Anchor::Center);
+
+                auto particleFall = GameToolbox::particleFromString("37a-1a2a0.33a15a-90a92a0a0a59a0a0a-10a0a0a0a0a1a1a0a48a1a0a0a0a0.27451a0a1a0a0a1a0a165a1a0a0.686275a0a0.235294a0a1a0a0a0a0.61a0a0a0a0a0a0a0a0a2a1a0a0a0a0a0a0a0a0a0a0a0a0a0a0a0a0a0a0", nullptr, false);
+                particleFall->setZOrder(-1);
+                particleFall->setScaleX(1.61f);
+                particleFall->setScaleY(1.61f);
+                supportSpr->addChildAtPosition(particleFall, Anchor::Bottom);
+
+                auto particleTop1 = GameToolbox::particleFromString("5a-1a2a0.33a15a-90a180a0a0a59a0a0a-10a0a0a0a0a20a1a75a48a1a0a0a0a0.27451a0a1a0a2a1a0a165a1a0a0.686275a0a0.235294a0a1a0a0a0a0.61a0a21a8a29a22a-11a40a1a2a1a0a0a0a158a0a0a0a0a0a0a0a0a0a0a0a0a0a0", nullptr, false);
+                particleTop1->setPosition({ supportSpr->getContentWidth() * .25f, supportSpr->getContentHeight() / 2 });
+                particleTop1->setZOrder(1);
+                particleTop1->setScaleX(.91f);
+                particleTop1->setScaleY(.91f);
+                particleTop1->setOpacity(105);
+                supportSpr->addChild(particleTop1);
+
+                auto particleTop2 = GameToolbox::particleFromString("5a-1a2a0.33a15a-90a180a0a0a59a0a0a-10a0a0a0a0a20a1a75a48a1a0a0a0a0.27451a0a1a0a2a1a0a165a1a0a0.686275a0a0.235294a0a1a0a0a0a0.61a0a21a8a29a22a-11a40a1a2a1a0a0a0a158a0a0a0a0a0a0a0a0a0a0a0a0a0a0", nullptr, false);
+                particleTop2->setPosition({ supportSpr->getContentWidth() * .75f, supportSpr->getContentHeight() / 2 });
+                particleTop2->setZOrder(1);
+                particleTop2->setScaleX(.91f);
+                particleTop2->setScaleY(.91f);
+                particleTop2->setOpacity(105);
+                supportSpr->addChild(particleTop2);
+
+                supportSpr->setScale(.8f);
+                supportSpr->runAction(CCRepeat::create(
+                    CCSequence::create(
+                        CCEaseInOut::create(CCScaleTo::create(1.5f, .7f), 4.f),
+                        CCEaseInOut::create(CCScaleTo::create(1.5f, .8f), 4.f),
+                        nullptr
+                    ),
+                    1000
+                ));
+                
+                auto supportBtn = CCMenuItemSpriteExtra::create(
+                    supportSpr, this, menu_selector(AboutBEPauseLayer::onSupport)
+                );
+                supportBtn->setID("support-be-btn"_spr);
+                topMenu->addChild(supportBtn);
+                topMenu->updateLayout();
+            }
+        }
 
         auto menu = this->getChildByID("guidelines-menu");
 
@@ -72,6 +132,9 @@ class $modify(AboutBEPauseLayer, EditorPauseLayer) {
         ));
     }
 
+    void onSupport(CCObject*) {
+        SupportPopup::create(true)->show();
+    }
     void onAbout(CCObject*) {
         AboutBEPopup::create()->show();
     }