+++ src/utils/Editor.hpp
@@ -1,0 +1,7 @@
+#pragma once
+
+#include <Geode/binding/EditorUI.hpp>
+
+using namespace geode::prelude;
+
+std::vector<GameObject*> getSelectedObjects(EditorUI* ui);